Add exceptions for language URL Generator
The language URL generator sometimes breaks some URLs that are
meant to switch between two different languages (for example
linking to a Dutch entity from a translation overview that's
viewed in English). We add a bunch of exceptions of targets and
origins where we disable the overwriting of links to combat this
issue.